+++ src/liquid/local-liquid-variables.js
