+++ src/liquid/vue-liquid-plugin.js
