+++ src/main.js
@@ -3,8 +3,8 @@ import Vue from 'vue';
 import App from './App.vue';
 import store from './store';
 import i18n from './i18n';
-import LiquidPlugin from './vue-liquid-plugin';
-import liquidVariables from './liquid-variables';
+import LiquidPlugin from './liquid/vue-liquid-plugin';
+import liquidVariables from './liquid/local-liquid-variables';
 import './vee-validate-config';
 import './vue-fontawesome-config';
 import 'bootstrap';
